+++ packages/stepper/src/Step.tsx
@@ -1,8 +1,7 @@
 import { useEffect, useRef, MutableRefObject, PropsWithChildren } from 'react';
 import { clsx } from '@equinor/fusion-react-styles';
 import { useStyles } from './style';
-import { Icon } from '@equinor/eds-core-react';
-import { done as doneStepIcon } from '@equinor/eds-icons';
+import { Badge } from './StepBadge';
 
 export type StepProps = {
   title: string;
@@ -16,32 +15,6 @@ export type StepProps = {
   done?: boolean;
   stepPaneRef?: MutableRefObject<HTMLElement>;
   stepCount?: number;
-};
-
-export type BadgeProps = {
-  position?: number;
-  active?: boolean;
-  done?: boolean;
-  runChange?: () => void;
-};
-
-const Badge = ({ position, active, done, runChange }: BadgeProps): JSX.Element => {
-  const styles = useStyles();
-  const badgeRef = useRef<HTMLDivElement>(null);
-
-  const badgeClasses = clsx(styles.badge, active && styles.active, done && styles.done);
-
-  useEffect(() => {
-    if (badgeRef.current && active) {
-      runChange && runChange();
-    }
-  }, [done]);
-
-  return (
-    <div ref={badgeRef} className={badgeClasses}>
-      {done ? <Icon data={doneStepIcon} /> : position}
-    </div>
-  );
 };
 
 export const Step = ({
@@ -96,6 +69,8 @@ export const Step = ({
   }
 
   return (
+    // TODO!
+    // eslint-disable-next-line jsx-a11y/anchor-is-valid, jsx-a11y/click-events-have-key-events, jsx-a11y/no-static-element-interactions
     <a onClick={() => !disabled && isClickable && onChange && onChange()} ref={stepRef} className={stepClasses}>
       <Badge position={position} active={isCurrent} done={done} runChange={() => onChange && onChange()} />
       <div className={styles.content}>
