+++ packages/stepper/src/index.tsx
@@ -1,2 +1,3 @@
-export { Step, BadgeProps, StepProps } from './Step';
+export { Step, StepProps } from './Step';
+export { type BadgeProps } from './StepBadge';
 export { Stepper, StepperProps } from './Stepper';
